requirements.txt
numpy
requests
opencv-python
moviepy
pygame
Pillow==9.5.0
pygetwindow
